Create host_enumeration.py
This is the first super handy Python script I have posted here. The script runs a variety of commands against a target host, then prints the output to a log file. I also like this script because it uses a prompt to ask the user for input and the password is kept hidden from stdout by way of the maskpass import. Cool stuff!